🚚 api: Move certificationAttestation model to certification/results context
