+++ api/lib/domain/models/index.js
@@ -1,7 +1,7 @@
 import { ComplementaryCertification } from '../../../src/certification/complementary-certification/domain/models/ComplementaryCertification.js';
-import { CertificationAttestation } from '../../../src/certification/course/domain/models/CertificationAttestation.js';
 import { AssessmentSimulator } from '../../../src/certification/flash-certification/domain/models/AssessmentSimulator.js';
 import { FlashAssessmentAlgorithm } from '../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
+import { CertificationAttestation } from '../../../src/certification/results/domain/models/CertificationAttestation.js';
 import { CertificationAssessmentScore } from '../../../src/certification/scoring/domain/models/CertificationAssessmentScore.js';
 import { CertificationCandidateForSupervising } from '../../../src/certification/session-management/domain/models/CertificationCandidateForSupervising.js';
 import { CertificationOfficer } from '../../../src/certification/session-management/domain/models/CertificationOfficer.js';
