+++ api/src/certification/results/domain/models/CertificationAttestation.js
